Issue #101: factored out code from ImportGeneOntology into OntologyImporter; drafted importer for CLO; added bax_derived_clo.ttl.
